feat: config.rs に設定をまとめる
