+++ src/history.rs
@@ -3,7 +3,6 @@ use crate::command::display::ecmascript::ECMAScriptStyle;
 use crate::command::display::lazy_k::LazyKStyle;
 use crate::command::Command;
 use crate::context::Context;
-use crate::display_style::DisplayStyle;
 use crate::parser::command::ecmascript::parse_command as parse_ecmascript_style_command;
 use crate::parser::command::lazy_k::parse_command as parse_lazy_k_style_command;
 use glob::glob;
@@ -13,6 +12,7 @@ use std::fs::File;
 use std::io::{self, BufRead, Write};
 use std::path::Path;
 use ulid::Ulid;
+use crate::config::{display_style, DisplayStyle};
 
 pub fn open_or_create_history_file() -> File {
     let dir = Path::new("~/.tuber")
@@ -66,7 +66,7 @@ pub struct Logger<W: Write>(W, DisplayStyle);
 
 impl<W: Write> Logger<W> {
     pub fn new(writer: W) -> Self {
-        Logger(writer, DisplayStyle::get())
+        Logger(writer, display_style())
     }
 
     pub fn push(&mut self, command: &Command) {
