+++ src/lib.rs
@@ -1,8 +1,8 @@
 pub mod command;
 pub mod context;
-pub mod display_style;
 pub mod evaluate;
 pub mod expression;
 pub mod function;
 pub mod identifier;
 pub mod parser;
+pub mod config;
